requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
django-ipware==7.0.1
django-rest-authemail==2.1.7
djangorestframework==3.15.1
idna==3.7
Markdown==3.6
pillow==10.3.0
python-ipware==3.0.0
requests==2.32.2
shortuuid==1.0.13
sqlparse==0.5.0
urllib3==2.2.1
